rework media handling
